Migrate all frontend components to new module
