+++ packages/protocol/tsconfig.json
@@ -6,5 +6,5 @@
     "declarationMap": false,
   },
   "include": ["src"],
-  "exclude": ["node_modules", "dist", "src/diff-type-checks.ts"],
+  "exclude": ["node_modules", "dist"],
 }
