+++ packages/toolkit-frontend/CHANGELOG.md
@@ -1,0 +1,44 @@
+# @arcanejs/diff
+
+## 0.3.1
+
+### Patch Changes
+
+- adfbcbf: Flesh out documentation for packages
+
+## 0.3.0
+
+### Minor Changes
+
+- fe0d7fb: Introduce advanced array diffing + patching (splices and nested changes)
+- 6424c0a: Add initial patch implementation
+
+## 0.2.0
+
+### Minor Changes
+
+- 779fae0: Initial implementation of diff
+
+## 0.1.4
+
+### Patch Changes
+
+- bumping to release new package
+
+## 0.1.3
+
+### Patch Changes
+
+- bump version again to publish tags
+
+## 0.1.2
+
+### Patch Changes
+
+- Bump to test tag publishing
+
+## 0.1.1
+
+### Patch Changes
+
+- 459ff7b: Initial publish
