+++ packages/toolkit-frontend/package.json
@@ -1,0 +1,69 @@
+{
+  "name": "@arcanejs/toolkit-frontend",
+  "version": "0.0.1",
+  "private": true,
+  "description": "Library of the frontend react components used to render the @arcanejs Toolkit",
+  "keywords": [],
+  "license": "MIT",
+  "author": {
+    "name": "Sam Lanning",
+    "url": "https://samlanning.com"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/arcanejs/arcanejs.git"
+  },
+  "exports": {
+    "./components": {
+      "@arcanejs/source": "./src/components/index.tsx",
+      "import": "./dist/components/index.mjs",
+      "require": "./dist/components/index.js",
+      "types": "./dist/components/index.d.ts"
+    },
+    "./components/core": {
+      "@arcanejs/source": "./src/components/core/index.ts",
+      "import": "./dist/components/core/index.mjs",
+      "require": "./dist/components/core/index.js",
+      "types": "./dist/components/core/index.d.ts"
+    },
+    "./util": {
+      "@arcanejs/source": "./src/util/index.ts",
+      "import": "./dist/util/index.mjs",
+      "require": "./dist/util/index.js",
+      "types": "./dist/util/index.d.ts"
+    },
+    "./styling": {
+      "@arcanejs/source": "./src/styling.ts",
+      "import": "./dist/styling.mjs",
+      "require": "./dist/styling.js",
+      "types": "./dist/styling.d.ts"
+    }
+  },
+  "scripts": {
+    "lint": "eslint . --max-warnings 0",
+    "build": "tsup"
+  },
+  "devDependencies": {
+    "@arcanejs/eslint-config": "workspace:^",
+    "@arcanejs/typescript-config": "workspace:^",
+    "@types/eslint": "^8.56.5",
+    "@types/node": "^20.11.24",
+    "@types/react": "^18",
+    "@types/react-dom": "^18",
+    "eslint": "^8.57.0",
+    "tsup": "^8.1.0",
+    "typescript": "^5.3.3"
+  },
+  "dependencies": {
+    "@arcanejs/protocol": "workspace:^",
+    "react": "^18",
+    "react-dom": "18.3.1",
+    "styled-components": "^6.1.13"
+  },
+  "files": [
+    "dist"
+  ],
+  "publishConfig": {
+    "access": "public"
+  }
+}
