+++ packages/toolkit-frontend/src/components/core/index.ts
@@ -1,0 +1,1 @@
+export { Icon } from './icon';
