+++ packages/toolkit-frontend/src/components/text-input.tsx
@@ -2,7 +2,9 @@ import * as React from 'react';
 import { styled } from 'styled-components';
 
 import * as proto from '@arcanejs/protocol';
+
 import { StageContext } from './context';
+
 import { THEME } from '../styling';
 
 interface Props {
