+++ packages/toolkit-frontend/src/util/index.ts
@@ -1,3 +1,5 @@
+export * from './touch';
+
 // TODO: move this to a new core-frontend library
 export const calculateClass = (
   ...args: (string | undefined | null | false)[]
