+++ packages/toolkit-frontend/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "@arcanejs/typescript-config/react-library.json",
+  "compilerOptions": {
+    "rootDir": "src",
+    "outDir": "dist"
+  },
+  "include": ["src"],
+  "exclude": ["node_modules", "dist"],
+}
