+++ packages/toolkit-frontend/tsconfig.lint.json
@@ -1,0 +1,8 @@
+{
+  "extends": "@arcanejs/typescript-config/base.json",
+  "compilerOptions": {
+    "outDir": "dist"
+  },
+  "include": ["src", "turbo",  "tsup.config.ts"],
+  "exclude": ["node_modules", "dist"]
+}
